% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeFindChromosomeInGroup.R
\name{findChromosomeInGroup}
\alias{findChromosomeInGroup}
\title{Function finds chromosome in group.}
\usage{
findChromosomeInGroup(dataFrame, chromosomeNumberOne, chromosomeNumberTwo)
}
\arguments{
\item{dataFrame}{gets the datframe of group dataframe}

\item{chromosomeNumberOne}{number of reverse Chromosome given by the user}

\item{chromosomeNumberTwo}{number of forward Chromosome given by the user}
}
\value{
result with group of chromosome
}
\description{
Function finds chromosome in group.
}
