% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeFindDrClusters.R
\name{findDrClusters}
\alias{findDrClusters}
\title{Function looks for discordant reads with the same chromosome name, mate and direction in close distance and assigna them the same id.}
\usage{
findDrClusters(configuredDf, tresholdValue, countBufferValue,
  numberOfPositionBases, numberOfMrnmBases, startIndex = NULL,
  endIndex = NULL, filteredChromomes = NULL)
}
\arguments{
\item{configuredDf}{gets dataframe result of function getConfiguration}

\item{tresholdValue}{specifies how many similar elements should be in the sequence}

\item{countBufferValue}{defines how many mismatched reads can be from another sequence to start ignore current sequence}

\item{numberOfPositionBases}{defines the maximum length of sequence position from the first element in sequence}

\item{numberOfMrnmBases}{defines the maximum length of sequence mrnm from the first element in sequence}

\item{startIndex}{defines from which position start read configured dataframe}

\item{endIndex}{defines to which position start read configured data from}

\item{filteredChromomes}{defines character vector of given chromosomes}
}
\value{
dataframe of sequences with the same id
}
\description{
Function looks for discordant reads with the same chromosome name, mate and direction in close distance and assigna them the same id.
}
