% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeGroupFiles.R
\name{groupFiles}
\alias{groupFiles}
\title{Function finds the groups of similar reads.}
\usage{
groupFiles(dataFrame, positionInPairDistance, sameDirection,
  distanceBetweenElements)
}
\arguments{
\item{dataFrame}{gets the datframe of similar sequences}
}
\value{
dataframe with groups

positionInPairDistance - distance between groups

sameDirection - decides if reads should be in the same diretion or not

distanceBetweenElements determines how many elements is between the center of a pair
}
\description{
Function finds the groups of similar reads.
}
