% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeConfiguration.R
\name{prepareConfiguration}
\alias{prepareConfiguration}
\title{Function prepares bam file to further processing}
\usage{
prepareConfiguration(file)
}
\arguments{
\item{file}{bam file}
}
\value{
dataframe with configured columns
}
\description{
Function prepares bam file to further processing
}
